pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "edu-ubb-gpu"
version = "0.1.0"

classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python"
]
description = "Create an ML model training job that runs on the EduML cluster"
license = { file = "LICENSE" }
authors = [
  {name = "Alexandru Manole", email = "alexandru.manole@ubbcluj.ro"}
]
maintainers = [
  {name = "Alexandru Manole", email = "alexandru.manole@ubbcluj.ro"}
]

requires-python = ">= 3.11"
dependencies = [
    "python-dotenv==1.0.1",
    "minio==7.2.10",
    "mlflow==2.17.2",
    "numpy==2.1.3",
    "pandas==2.2.3",
    "scikit-learn==1.5.2",
]
